+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Finished the base version of the `getInterfaceID` function.
 * Used `Hardhat` and `solidity-ast` packages to parse the contract and get the interface ID.
+* Added support for passing contract names to the `getInterfaceID` function.